Simplify compose view controller
